Refactor + remove white spaces
